+++ Runtime/Scripts/Export/ExportJobs.cs
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2023 Unity Technologies and the glTFast authors
 // SPDX-License-Identifier: Apache-2.0
 
+using GLTFast.Jobs;
 using Unity.Burst;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
@@ -60,7 +61,8 @@ namespace GLTFast.Export
         public unsafe struct ConvertPositionFloatJob : IJobParallelFor
         {
 
-            public uint byteStride;
+            public uint inputByteStride;
+            public uint outputByteStride;
 
             [ReadOnly]
             [NativeDisableUnsafePtrRestriction]
@@ -72,8 +74,8 @@ namespace GLTFast.Export
 
             public void Execute(int i)
             {
-                var inPtr = (float3*)(input + i * byteStride);
-                var outPtr = (float3*)(output + i * byteStride);
+                var inPtr = (float3*)(input + i * inputByteStride);
+                var outPtr = (float3*)(output + i * outputByteStride);
 
                 var tmp = *inPtr;
                 tmp.x *= -1;
@@ -85,7 +87,8 @@ namespace GLTFast.Export
         public unsafe struct ConvertTangentFloatJob : IJobParallelFor
         {
 
-            public uint byteStride;
+            public uint inputByteStride;
+            public uint outputByteStride;
 
             [ReadOnly]
             [NativeDisableUnsafePtrRestriction]
@@ -97,8 +100,8 @@ namespace GLTFast.Export
 
             public void Execute(int i)
             {
-                var inPtr = (float4*)(input + i * byteStride);
-                var outPtr = (float4*)(output + i * byteStride);
+                var inPtr = (float4*)(input + i * inputByteStride);
+                var outPtr = (float4*)(output + i * outputByteStride);
 
                 var tmp = *inPtr;
                 tmp.z *= -1;
@@ -109,7 +112,8 @@ namespace GLTFast.Export
         [BurstCompile]
         public unsafe struct ConvertTexCoordFloatJob : IJobParallelFor
         {
-            public uint byteStride;
+            public uint inputByteStride;
+            public uint outputByteStride;
 
             [ReadOnly]
             [NativeDisableUnsafePtrRestriction]
@@ -121,12 +125,36 @@ namespace GLTFast.Export
 
             public void Execute(int i)
             {
-                var inPtr = (float2*)(input + i * byteStride);
-                var outPtr = (float2*)(output + i * byteStride);
+                var inPtr = (float2*)(input + i * inputByteStride);
+                var outPtr = (float2*)(output + i * outputByteStride);
 
                 var tmp = *inPtr;
                 tmp.y = 1 - tmp.y;
                 *outPtr = tmp;
+            }
+        }
+
+        [BurstCompile]
+        public unsafe struct ConvertGenericJob : IJobParallelFor
+        {
+            public uint inputByteStride;
+            public uint outputByteStride;
+
+            public uint byteLength;
+
+            [ReadOnly]
+            [NativeDisableUnsafePtrRestriction]
+            public byte* input;
+
+            [WriteOnly]
+            [NativeDisableUnsafePtrRestriction]
+            public byte* output;
+
+            public void Execute(int i)
+            {
+                var inPtr = input + i * inputByteStride;
+                var outPtr = output + i * outputByteStride;
+                UnsafeUtility.MemCpy(outPtr, inPtr, byteLength);
             }
         }
     }