+++ Samples/Documentation/Manual/ExtraData.cs
@@ -7,7 +7,7 @@ namespace Samples.Documentation.Manual
     #region ExtraData
     using UnityEngine;
 
-    public class ExtraData : MonoBehaviour
+    class ExtraData : MonoBehaviour
     {
         public string someExtraKey;
     }