+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Faster buffer conversion jobs due to batching via [`IJobParallelForBatch`](https://docs.unity3d.com/Packages/com.unity.collections@2.4/api/Unity.Jobs.IJobParallelForBatch.html).
+- (Export) Vertex attributes are discarded if they are not used/referenced.
+
+### Fixed
+- (Export) Discrepancy in color due to export of unused vertex colors.
 
 ### Removed
 - Soft dependency on deprecated [Unity Jobs package][JobsPkg].